Filters

See merge request mont_int_2022/int19!25

+++ backend/src/main/java/internship/paymentSystem/backend/models/Balance.java
@@ -5,6 +5,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Date;
 
@@ -20,10 +21,10 @@ public class Balance {
     private Long id;
 
     @Column(name = "total")
-    private Double total;
+    private BigDecimal total;
 
     @Column(name = "available")
-    private Double available;
+    private BigDecimal available;
 
     @Column(name = "account_id")
     private Long accountID;
@@ -33,7 +34,7 @@ public class Balance {
     @CreationTimestamp
     private LocalDateTime timestamp;
 
-    public Balance(Double total, Double available, Long accountID) {
+    public Balance(BigDecimal total, BigDecimal available, Long accountID) {
         this.total = total;
         this.available = available;
         this.accountID = accountID;
